medical-chart-tagger/requirements.txt
doccano==1.8.2
pandas==1.5.3
numpy==1.23.5
spacy==3.5.0
pyyaml==6.0